LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/reporting/access-reporting/property-postmessage-2.https-expected.txt

CONSOLE MESSAGE: Unable to post message to http://localhost:8800. Recipient has origin https://localhost:9443.


Harness Error (TIMEOUT), message = null

TIMEOUT same-origin > w => w.postMessage("") Test timed out
NOTRUN cross-origin > w => w.postMessage("")
